Mostly fixed the mouse movement in TMWserv client.
The only case left in mouse movement is that the client doesn't check
for walkability of the destination point when you click on a unwalkable point.

Let me explain this NASTY bug:

The bug remained in the Being::SetDestination() function,
when recalculating the path from tiles to pixels.
The changeX and changeY variables went crazy when (endX - startX) gave a negative value.
That why the given path became random, and the player went anywhere.

This didn't happened to monsters, NPCs, and when walking using the keyboard
because the patnodes system isn't used for movement up to 1 tile at a time.

I removed some dead code (in viewpoint.cpp), made the keyboard navigation a little
bit more bearable, and fixed this client bug.

Now, I'll be up to finish fixing the movement system using mouse (What's remaining is a joke
next to what I had to do to discover this...),
and I'll look at a third time to the keyboard system which is a bit raw, just for now...

Regards,

P.S.: Kage, I'd like to get a three-cheese pizza!

+++ src/being.cpp
@@ -168,16 +168,16 @@ void Being::setDestination(int dstX, int dstY)
     // Interpolate the offsets.  Also convert from tile based to pixel based
 
     // Find the starting offset
-    int startX = (srcX % 32);
-    int startY = (srcY % 32);
+    float startX = (srcX % 32);
+    float startY = (srcY % 32);
 
     // Find the ending offset
-    int endX = (dstX % 32);
-    int endY = (dstY % 32);
+    float endX = (dstX % 32);
+    float endY = (dstY % 32);
 
     // Find the distance, and divide it by the number of steps
-    int changeX = (endX - startX) / thisPath.size();
-    int changeY = (endY - startY) / thisPath.size();
+    int changeX = (int)((endX - startX) / thisPath.size());
+    int changeY = (int)((endY - startY) / thisPath.size());
 
     // Convert the map path to pixels over tiles
     // And add interpolation between the starting and ending offsets
